LayoutTests/platform/chromium-mac/fast/canvas/webgl/texImageTest-expected.txt

Test of texImage2d and texSubImage2d

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.texImage2D(context.TEXTURE_2D) threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS context.texImage2D(context.TEXTURE_2D, 0, context.RGBA, 64, 64, 0, context.RGBA, context.UNSIGNED_BYTE, null) is undefined.
PASS context.texImage2D(context.TEXTURE_2D, 0, context.RGBA, 0, context.RGBA, context.UNSIGNED_BYTE, 0) threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS context.texImage2D(context.TEXTURE_2D, 0, context.RGBA, 2, 2, 0, context.RGBA, context.UNSIGNED_BYTE, array) is undefined.
PASS context.texImage2D(context.TEXTURE_2D, 0, imageData, true) is undefined.
PASS context.texImage2D(context.TEXTURE_2D, 0, image) is undefined.
PASS context.texImage2D(context.TEXTURE_2D, 0, canvas2d, false) is undefined.
PASS context.texImage2D(context.TEXTURE_2D, 0, video, true, false) is undefined.
PASS context.texSubImage2D(context.TEXTURE_2D) threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, 0, 0, context.RGBA, context.UNSIGNED_BYTE, null) is undefined.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, 0, 0, context.RGBA, context.UNSIGNED_BYTE, 0) threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, 0, context.UNSIGNED_BYTE, 0) threw exception Error: SYNTAX_ERR: DOM Exception 12.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, imageData, false) is undefined.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, image) is undefined.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, canvas2d, true) is undefined.
PASS context.texSubImage2D(context.TEXTURE_2D, 0, 10, 20, video, false, true) is undefined.
PASS successfullyParsed is true

TEST COMPLETE
